chore: 🤖 navbar button integrations
