+++ src/pages/index.astro
@@ -13,7 +13,7 @@ import Footer from '@components/Footer';
 ---
 
 <LandingPageHtml>
-  <Navbar pathname={Astro.url.pathname} client:load />
+  <Navbar pathname={Astro.url.pathname} client:only="react" />
   <Hero />
   <Partners />
   <DeployOnFleek />
